requirements.txt
openai
python-dotenv
fastapi
uvicorn
websockets
jinja2
python-multipart